Update eulertotient
The Euler totient of a number n is the number of integers between 1 and n that are relatively prime to n. By convention, the totient of 1 is 1. I define a function f that takes an argument m and returns a vector of the divisors of m. Two numbers m and n are relatively prime if (f m) and (rest (f n)) have no elements in common (1 is a divisor of every integer.)