Create wordchains
This is problem 82 on 4clojure. Given a set of words, this function returns true if they can be arranged into a word chain, false otherwise.

We first define a function wordpair, which tests whether two words can be adjacent in a word chain. Then we build longer and longer word chains by concatenating existing chains (making sure not to double back). If this process terminates (ie if result and storage are equal) and none of these chains are as long as the original set, then we return false. If we do have a long enough chain, then we return true.